requirements.txt
p2pnetwork==1.1
pycryptodome==3.12.0